fully working. need to improve ui & bug squash
* server.js is the pdf processor & server (using express)
* main.js & main.html are client side interfaces.
* when deploying to server, make sure to change request post link in main.html and verify that it works with sever.js
