--- index.html
@@ -1,29 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>CPS Schedule Generator</title>
-    <!-- <script src="bundle.js" type="text/javascript"></script> -->
-    <script src="scripts/index.js"></script>
-</head>
-
-<body>
-
-    <!-- <button onclick="start()"></button> -->
-    <form action="javascript:;" onsubmit="starting()">
-        First Class: <input type="text" id="a"><br>
-        Second Class: <input type="text" id="b"><br>
-        Third Class: <input type="text" id="c"><br>
-        Fourth Class: <input type="text" id="d"><br>
-        Fifth Class: <input type="text" id="e"><br>
-        Sixth Class: <input type="text" id="f"><br>
-        Seventh Class: <input type="text" id="g"><br>
-        <input type="submit" value="Submit">
-    </form>
-    <h3 id="schGenerating"></h3>
-</body>
-
-</html>
