+++ main.html
@@ -1,0 +1,42 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title>CPS Schedule Generator</title>
+    <!-- <script src="bundle.js" type="text/javascript"></script> -->
+    <!-- <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.10.2/jquery.min.js"></script> -->
+    <script src="main.js"></script>
+</head>
+
+<body>
+
+    <!-- <button onclick="start()"></button> -->
+    <script src="main.js"></script>
+    <form id="classForm" method="POST" action="http://localhost:8000/">
+        First Class: <input type="text" id="a" name="first"><br>
+        Second Class: <input type="text" id="b" name="second"><br>
+        Third Class: <input type="text" id="c" name="third"><br>
+        Fourth Class: <input type="text" id="d" name="fourth"><br>
+        Fifth Class: <input type="text" id="e" name="fifth"><br>
+        Sixth Class: <input type="text" id="f" name="sixth"><br>
+        Seventh Class: <input type="text" id="g" name="seventh"><br>
+        <!-- <label for="name">Name:</label>
+    <input type="text" id="name" name="name" placeholder="Enter your full name" />
+
+    <label for="email">Email:</label>
+    <input type="email" id="email" placeholder="Enter your email address" />
+
+    <label for="message">Message:</label>
+    <textarea id="message" placeholder="What's on your mind?"></textarea>
+
+    <input type="submit" value="Send message" /> -->
+        <input type="submit" value="Submit">
+    </form>
+    <!-- <script src="main.js"></script> -->
+    <h3 id="schGenerating"></h3>
+</body>
+
+</html>
