--- scripts/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "-cps-schedule-generator",
-  "version": "1.0.0",
-  "description": "CPS Schedule Generator -> Auto-fill PDF",
-  "main": "index.js",
-  "dependencies": {
-    "@pdf-lib/fontkit": "0.0.4",
-    "browserify-fs": "^1.0.0",
-    "connect": "^3.7.0",
-    "node-fetch": "^2.6.0",
-    "pdf-lib": "^1.2.1",
-    "serve-static": "^1.14.1"
-  },
-  "browser": {
-    "fs": "fs"
-  },
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/AirFusion45/-CPS-Schedule-Generator.git"
-  },
-  "author": "AirFusion45",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/AirFusion45/-CPS-Schedule-Generator/issues"
-  },
-  "homepage": "https://github.com/AirFusion45/-CPS-Schedule-Generator#readme"
-}
